+++ tests/MessageBuilders/TextMessageBuilderTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Phine\Tests\Helpers\MessageBuilders;
+namespace Phine\Tests\MessageBuilders;
 
 use LINE\Clients\MessagingApi\Model\Emoji;
 use LINE\Constants\MessageType;
-use Phine\Helpers\MessageBuilders\TextMessageBuilder;
+use Phine\MessageBuilders\TextMessageBuilder;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
